Renamed main entrypoint
