+++ src/index.html
@@ -4,7 +4,7 @@
     <title>Imagelightbox</title>
     <meta charset='utf-8'>
     <meta http-equiv="X-UA-Compatible" content="chrome=1">
-    <script type="module" src="./lib/imagelightbox.ts"></script>
+    <script type="module" src="./lib/imagelightbox.jquery.ts"></script>
     <script type="module" src="./index.ts"></script>
     <link href='https://fonts.googleapis.com/css?family=Chivo:900' rel='stylesheet' type='text/css'>
     <link rel="stylesheet" type="text/css" href="./index.css">
